timer rennder 추가

+++ include/math/Vector2.hpp
@@ -1,11 +1,11 @@
 #pragma once
-
+#include <inttypes.h>
 namespace neo::math
 {
 	struct Vector2
 	{
-		Vector2():_y(0.f),_x(0.f){}
-		Vector2(const float& x,const float& y):_x(x),_y(y){}
+		Vector2():_y(0),_x(0){}
+		Vector2(const int32_t& x,const int32_t& y):_x(x),_y(y){}
 		Vector2(const Vector2& vec2)
 		{
 			this->_x = vec2._x;
@@ -22,7 +22,7 @@ namespace neo::math
 			return Vector2( this->_x - vec2._x, this->_y - vec2._y );
 		}
 
-		float _x = 0.f;
-		float _y = 0.f;
+		int32_t _x = 0;
+        int32_t _y = 0;
 	};
 }
